fix: lint, useless files and separate update from auth flow
