+++ back/tests/__helpers__/context.ts
@@ -8,14 +8,12 @@ import { PrismaClient } from '@prisma/client';
 import axios, { AxiosInstance } from 'axios';
 
 import Server from '../../src/Server';
-import { handleRequestFailure } from './async';
 
 interface TestServer {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   graphql: (query: DocumentNode, variables?: unknown) => Promise<any>;
   stop: () => void;
   baseURL: string;
-  port: number;
 }
 
 export interface TestContext {
@@ -46,7 +44,6 @@ function startTestServer(): TestServer {
     stop,
     graphql,
     baseURL,
-    port: (httpServer.address() as AddressInfo).port,
   };
 }
 
@@ -59,7 +56,6 @@ export function createTestContext(): TestContext {
     ctx.server.graphql = testServer.graphql;
     ctx.server.stop = testServer.stop;
     ctx.server.baseURL = testServer.baseURL;
-    ctx.server.port = testServer.port;
   });
 
   afterAll(async () => {
@@ -70,8 +66,5 @@ export function createTestContext(): TestContext {
 }
 
 export function createAxiosInstance(ctx: TestContext): AxiosInstance {
-  const instance = axios.create({ baseURL: ctx.server.baseURL });
-  // instance.interceptors.response.use((res) => res, handleRequestFailure);
-
-  return instance;
+  return axios.create({ baseURL: ctx.server.baseURL });
 }
