--- back/tests/__helpers__/generate.ts
@@ -1,2 +1,0 @@
-import faker from 'faker';
-
