--- back/tests/data/testLib/c3pm.yml
@@ -1,16 +1,0 @@
-c3pm_version: v1
-type: executable
-name: testLib
-description: testLib project
-version: 2.0.0
-standard: "20"
-license: MIT
-files:
-  sources:
-  - '**/*.cpp'
-  includes:
-  - '**/*.hpp'
-  include_dirs: []
-  exported_dir: ""
-  exported_include_dirs: []
-dependencies: {}
