--- back/tests/data/testLib/src/main.cpp
@@ -1,6 +1,0 @@
-#include <iostream>
-
-int main() {
-	std::cout << "Hello c3pm!" << std::endl;
-	return 0;
-}
